fix: fix register
