+++ packages/frontend/src/app/(sticky-header)/register/page.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next'
 import Link from 'next/link'
 import { GENERAL_DESCRIPTION } from '@/app/constants'
+import { LoginBtns } from '@/app/components/login'
 // import { isProduction } from '@/environment-variables'
 
 export const metadata: Metadata = {
@@ -9,20 +10,12 @@ export const metadata: Metadata = {
 }
 
 // export const revalidate = isProduction ? 86400 : 0 // 1 day
-
-export default async function Login() {
+export default async function Register() {
   return (
-    <main className="flex flex-col justify-center items-center gap-10 bg-gray-100">
+    <main className="flex flex-col justify-center items-center gap-10">
       <span>註冊</span>
       <span>{'免費獲得電子報儲存喜愛的深度報導'}</span>
-      <div id="team" style={{ width: '95vw', scrollMarginTop: '62px' }}></div>
-      <div
-        id="consultants"
-        style={{ width: '95vw', scrollMarginTop: '62px' }}
-      ></div>
-      <div className="border p-3">{'使用Google帳號'}</div>
-      <div className="border p-3">{'使用Facebook帳號'}</div>
-      <div className="border p-3">{'使用電子信箱'}</div>
+      <LoginBtns />
       <Link href={'/login'}>{'已經有帳號了，我要登入'}</Link>
       <span>{'完成帳號登入代表你同意我們的隱私權政策'}</span>
     </main>
